Renamed useRegions to useLoadedRegions
And removed requirement to pass in ReplayClient
