+++ packages/replay-next/components/console/Focuser.tsx
@@ -1,11 +1,10 @@
 import { TimeStampedPointRange } from "@replayio/protocol";
-import React, { MutableRefObject, useContext, useEffect, useRef, useState } from "react";
+import { MutableRefObject, useContext, useEffect, useRef, useState } from "react";
 
 import { FocusContext } from "replay-next/src/contexts/FocusContext";
 import { SessionContext } from "replay-next/src/contexts/SessionContext";
-import useLoadedRegions from "replay-next/src/hooks/useRegions";
+import useLoadedRegions from "replay-next/src/hooks/useLoadedRegions";
 import { formatTimestamp } from "replay-next/src/utils/time";
-import { ReplayClientContext } from "shared/client/ReplayClientContext";
 
 import styles from "./Focuser.module.css";
 
@@ -57,8 +56,7 @@ function RangeSlider({
   end: number;
   onChange: (begin: number, end: number) => void;
 }) {
-  const client = useContext(ReplayClientContext);
-  const loadedRegions = useLoadedRegions(client);
+  const loadedRegions = useLoadedRegions();
 
   const ref = useRef<HTMLDivElement>(null) as MutableRefObject<HTMLDivElement>;
