+++ src/ui/components/Timeline/UnloadedRegions.tsx
@@ -2,13 +2,13 @@ import { TimeStampedPointRange } from "@replayio/protocol";
 import clamp from "lodash/clamp";
 import { FC } from "react";
 
-import { getLoadedRegions } from "ui/reducers/app";
+import useLoadedRegions from "replay-next/src/hooks/useLoadedRegions";
 import { getFocusWindow, getZoomRegion } from "ui/reducers/timeline";
 import { useAppSelector } from "ui/setup/hooks";
 import { getVisiblePosition, overlap } from "ui/utils/timeline";
 
 export const UnloadedRegions: FC = () => {
-  const loadedRegions = useAppSelector(getLoadedRegions);
+  const loadedRegions = useLoadedRegions();
   const zoomRegion = useAppSelector(getZoomRegion);
   const focusWindow = useAppSelector(getFocusWindow);
